MAJOR COMMIT: restructured main filetree.
