--- src/IconoWraper/IconoWraper.module.scss
@@ -1,3 +1,0 @@
-.iconoWraper{
-    background-color: black;
-}
