+++ src/componentes/EcosystemCard/EcosystemCard.module.scss
