+++ src/componentes/IconoWraper/IconoWraper.module.scss
@@ -1,7 +1,3 @@
-.iconoWrapper {
-    cursor: pointer;
-  
-    &:hover {
-      translate: ease 0.2;
-    }
-  }
+.iconoWraper{
+    background-color: black;
+}
